refactor: repository tests into their own util
