+++ test/src/firefuel_repository_test.dart
@@ -1,47 +1,32 @@
 import 'package:firefuel/firefuel.dart';
-import 'package:firefuel/src/collection.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../mocks/mock_collection.dart';
 import '../utils/expected_failure.dart';
 import '../utils/test_user.dart';
+import 'utils/repository_test_util.dart';
 
 void main() {
   final defaultUser = TestUser('testUser');
-  late MockCollection mockCollection;
-  late FirefuelRepository testRepository;
 
-  setUp(() {
-    mockCollection = MockCollection();
-    testRepository = TestRepository(collection: mockCollection);
-  });
-
-  group('#create', () {
-    test('should return Right when successful', () async {
+  RepositoryTestUtil.runTests(
+    methodName: 'create',
+    mockCollection: MockCollection(),
+    onHappyPath: (mockCollection, testRepository) async {
       mockCollection.initialize(
           onCreate: () => DocumentId('h34jfhg43fiuy3gv4'));
 
-      final createResult = await testRepository.create(value: defaultUser);
-
-      expect(createResult, isA<Right<Failure, DocumentId>>());
-    });
-
-    test('should return Left when not successful', () async {
+      return testRepository.create(value: defaultUser);
+    },
+    onSadPath: (mockCollection, testRepository) async {
       mockCollection.initialize(onCreate: () => throw ExpectedFailure());
 
-      final createResult = await testRepository.create(value: defaultUser);
-
-      expect(createResult, isA<Left<Failure, DocumentId>>());
-    });
-  });
+      return testRepository.create(value: defaultUser);
+    },
+  );
 
   group('#read', () {}, skip: true);
   group('#readAsStream', () {}, skip: true);
   group('#update', () {}, skip: true);
   group('#delete', () {}, skip: true);
-}
-
-class TestRepository<T extends Serializable> extends FirefuelRepository<T> {
-  TestRepository({required Collection<T> collection})
-      : super(collection: collection);
 }
